+++ packages/e2e-tests/package.json
@@ -26,6 +26,7 @@
 		"@wordpress/jest-console": "file:../jest-console",
 		"@wordpress/jest-puppeteer-axe": "file:../jest-puppeteer-axe",
 		"@wordpress/scripts": "file:../scripts",
+		"@wordpress/url": "file:../url",
 		"expect-puppeteer": "^4.3.0",
 		"lodash": "^4.17.15",
 		"uuid": "^3.3.2"
